Create Purchase Order and Supplier API.

+++ ERP/vitality-common/src/main/java/com/vitality/common/dtos/CreateSupplierResponse.java
@@ -3,11 +3,9 @@ package com.vitality.common.dtos;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigInteger;
-
 @Data
 @RequiredArgsConstructor
 public class CreateSupplierResponse {
-    private BigInteger supplierId;
+    private Long supplierId;
     private String supplierName;
 }
